add interfacesto dao package insted daoclasses and add @repository and convert Long field to int

+++ src/main/java/ir/maktab/homeservicesystem/data/dao/AddressDao.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AddressDao extends JpaRepository<Address, Long> {
+public interface AddressDao extends JpaRepository<Address, Integer> {
 }
